server/python_pdf_service/requirements.txt
flask==2.0.1
PyMuPDF==1.19.6
pdfminer.six==20221105
gunicorn==20.1.0 